requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
motor==3.4.0
pydantic==2.5.3
email-validator==2.1.0
pymongo==4.7.0
python-dotenv==1.0.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
requests==2.31.0

# Google API (Gmail, etc.)
google-auth-oauthlib
google-auth-httplib2
google-api-python-client

# TensorFlow for AI Model
tensorflow>=2.15.0
pillow>=10.0.0
numpy>=1.24.0

# Image Preprocessing for AI Model Domain Gap
rembg[cpu]>=2.0.50
opencv-python>=4.8.0
